Move permission parsing to BoxItem
