+++ box-content-sdk/src/main/java/com/box/androidsdk/content/models/BoxBookmark.java
@@ -42,8 +42,6 @@ public class BoxBookmark extends BoxItem {
             FIELD_COMMENT_COUNT
     };
 
-    protected transient EnumSet<Permission> mPermissions = null;
-
     /**
      * Constructs an empty BoxBookmark object.
      */
@@ -97,18 +95,6 @@ public class BoxBookmark extends BoxItem {
         return null;
     }
 
-    /**
-     * Gets the permissions that the current user has on the bookmark.
-     *
-     * @return the permissions that the current user has on the bookmark.
-     */
-    public EnumSet<Permission> getPermissions() {
-        if (mPermissions == null) {
-            parsePermissions();
-        }
-        return mPermissions;
-    }
-
     @Override
     protected void parseJSONMember(JsonObject.Member member) {
         String memberName = member.getName();
@@ -116,42 +102,7 @@ public class BoxBookmark extends BoxItem {
         if (memberName.equals(FIELD_URL)) {
             this.mProperties.put(FIELD_URL, value.asString());
             return;
-        } else if (memberName.equals(FIELD_PERMISSIONS)) {
-            BoxPermission permission = new BoxPermission();
-            permission.createFromJson(value.asObject());
-            this.mProperties.put(FIELD_PERMISSIONS, permission);
-            parsePermissions();
-            return;
         }
         super.parseJSONMember(member);
-    }
-
-
-    private EnumSet<Permission> parsePermissions() {
-        BoxPermission permission = (BoxPermission) this.mProperties.get(FIELD_PERMISSIONS);
-        if (permission == null)
-            return null;
-
-        Map<String, Object> permissionsMap = permission.getPropertiesAsHashMap();
-        mPermissions = EnumSet.noneOf(Permission.class);
-        for (Map.Entry<String, Object> entry : permissionsMap.entrySet()) {
-            // Skip adding all false permissions
-            if (entry.getValue() == null || !(Boolean) entry.getValue())
-                continue;
-
-            String key = entry.getKey();
-            if (key.equals(Permission.CAN_RENAME.toString())) {
-                mPermissions.add(Permission.CAN_RENAME);
-            } else if (key.equals(Permission.CAN_DELETE.toString())) {
-                mPermissions.add(Permission.CAN_DELETE);
-            } else if (key.equals(Permission.CAN_SHARE.toString())) {
-                mPermissions.add(Permission.CAN_SHARE);
-            } else if (key.equals(Permission.CAN_SET_SHARE_ACCESS.toString())) {
-                mPermissions.add(Permission.CAN_SET_SHARE_ACCESS);
-            } else if (key.equals(Permission.CAN_COMMENT.toString())) {
-                mPermissions.add(Permission.CAN_COMMENT);
-            }
-        }
-        return mPermissions;
     }
 }
